feat:added refresh tokens and fixed some things

+++ API/DTOs/FakturaDto.cs
@@ -6,12 +6,12 @@ namespace API.DTOs
         public int Broj { get; set; }
         public DateTime Datum { get; set; }
         public string Partner { get; set; }
-        public float IznosBezPdv { get; set; }
-        public float PostoRabata { get; set; }
-        public float Rabat { get; set; }
-        public float IznosSaRabatomBezPdv { get; set; }
-        public float Pdv { get; set; }
-        public float Ukupno { get; set; }
+        public double IznosBezPdv { get; set; }
+        public double PostoRabata { get; set; }
+        public double Rabat { get; set; }
+        public double IznosSaRabatomBezPdv { get; set; }
+        public double Pdv { get; set; }
+        public double Ukupno { get; set; }
         public List<StavkaFaktureDto> StavkeFakture { get; set; }
     }
 }
