+++ API/Entities/StavkaFakture.cs
@@ -9,13 +9,13 @@ namespace API.Entities
         public int Rbr { get; set; }
         public string NazivArtikla { get; set; }
         public int Kolicina { get; set; }
-        public float Cijena { get; set; }
-        public float IznosBezPdv { get; set; }
-        public float PostoRabata { get; set; }
-        public float Rabat { get; set; }
-        public float IznosSaRabatomBezPdv { get; set; }
-        public float Pdv { get; set; }
-        public float Ukupno { get; set; }
+        public double Cijena { get; set; }
+        public double IznosBezPdv { get; set; }
+        public double PostoRabata { get; set; }
+        public double Rabat { get; set; }
+        public double IznosSaRabatomBezPdv { get; set; }
+        public double Pdv { get; set; }
+        public double Ukupno { get; set; }
         public int FakturaId { get; set; }
         public Faktura Faktura { get; set; }
     }
